LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/timeline-scope.html

<!DOCTYPE html>
<title>Behavior of the timeline-scope property</title>
<link rel="help" src="https://github.com/w3c/csswg-drafts/issues/7759">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>

<main id=main></main>
<script>
  async function inflate(t, template) {
    t.add_cleanup(() => main.replaceChildren());
    return runAndWaitForFrameUpdate(() => {
       main.append(template.content.cloneNode(true));
    });
  }

  async function scrollTop(e, value) {
    e.scrollTop = value;
    await waitForNextFrame();
  }
</script>
<style>
  @keyframes anim {
    from { width: 100px; }
    to { width: 200px; }
  }

  .scroller {
    overflow-y: hidden;
    width: 200px;
    height: 200px;
  }
  .scroller > .content {
    margin: 400px 0px;
    width: 100px;
    height: 100px;
    background-color: green;
  }
  .target {
    background-color: coral;
    width: 100px;
    animation: anim auto linear;
    animation-timeline: --t1;
  }
  .timeline {
    scroll-timeline-name: --t1;
  }
  .scope {
    timeline-scope: --t1;
  }

</style>

<!-- Basic Behavior -->

<template id=deferred_timeline>
  <div class="scope">
    <div class=target>Test</div>
    <div class="scroller timeline">
      <div class=content></div>
    </div>
  </div>
</template>
<script>
  promise_test(async (t) => {
    await inflate(t, deferred_timeline);
    
    deferred_timeline.remove();
    
    let scroller = main.querySelector('.scroller');
    let target = main.querySelector('.target');

    const anim = target.getAnimations()[0];
    // console.log("Created animation");
    await anim.ready;

    console.log("Animation ready, about to scroll");
    await scrollTop(scroller, 350); // 50%
    // console.log("Done scrolling");

    // assert_equals(anim.currentTime.toString(), "50%");
    console.log("Getting computed style");
    const computedWidth = getComputedStyle(target).width;
    console.log("Got computed style");
    assert_equals(computedWidth, '150px'); // 100px => 200px, 50%
  }, 'Descendant can attach to deferred timeline');
</script>

</script>
